requirements.txt
numpy==1.19.0
pandas==1.2.0
matplotlib==3.4.0
streamlit==1.1.0
Pillow==8.2.0
scikit-learn==0.24.2